feat: add work history feature to home
I have added a feature that shows my work history in the home page. The functionality and way of getting data is the same as "PorjectsRow" component. It gets data from a Json database and right now it doesnt have the option to be linked with a href which should be added in future.