feat: add category carousel for column projects page
Changed the whole style and functionality of the projects page category. Now, each category has a icon related to the category.

Added carousel functionality using Swiper.js that lets the user to navigate through the categories more easily.

Also made the useSwipe to not consider the projects category carousel  swipes as the page swipe.